requirements.txt
jira==3.5.1
PyYAML==6.0.1
